more graphs and some API changes

+++ tornapi_code/process_web_api_changes.py
@@ -18,12 +18,14 @@ def check_and_store(apikey, faction):
     try:
         data = r.json()
     except:
-        return
+        return "FAIL"
     if 'player_id' in data:
         p_id = data['player_id']
         c2.execute("""delete from apikeys where player_id = ?""", (p_id,))
         c2.execute("""insert into apikeys(et,player_id,short_err,long_err,key) values(?,?,?,?,?)""", (int(time.time()), p_id, 0, 0, apikey,))
+        c2.execute("""update playerwatch set latest=latest/2  where player_id=?""", (p_id,))
         conn2.commit()
+        return "OK"
 
 def delete_api_key(p_id):
     c2.execute("""delete from apikeys where player_id = ?""", (p_id,))
@@ -45,27 +47,37 @@ def read_and_update(filename):
                 break
 
     if len(text) < 3:
+        print("\n unfinished ", filename, "\n")
         return # unfinished file
+
+    do_remove = False
 
     if len(text) < 5:
         if text[0] == 'APIKEY':
             if text[3] == 'END':
                 apikey = text[1]
                 faction = text[2]
-                check_and_store(apikey, faction)
+                print("\n calling check_and_store() ", filename, "\n")
+                cs_status = check_and_store(apikey, faction)
+                if "OK" == cs_status:
+                    do_remove = True
             else:
+                print("\nwrong content:", filename, text, "\n")
                 return 
         elif text[0] == 'DELETE APIKEY':
             if text[2] == 'END':
                 p_id = text[1]
+                print("\n deleting ...\n")
                 delete_api_key(p_id)
             else:
+                print("\nwrong content:", filename, text, "\n")
                 return 
         else:
             print('unexpected', text)
             return 
 
-    os.remove(filename)
+    if do_remove:
+        os.remove(filename)
 
 
 if __name__ == "__main__":
@@ -82,7 +94,9 @@ if __name__ == "__main__":
             if is_numeric:
                 print("planning to work on", entry.name)
                 read_and_update(entry.name)
+                print("finished work on", entry.name)
             else:
+                print("removing non-numeric filename", entry.name)
                 os.remove(entry.name)
 
     c2.close()
